requirements.txt
torch
torchvision
timm
open-clip-torch
pyyaml
git+https://github.com/facebookresearch/segment-anything.git
